+++ client/src/components/InputAndDeleteButton/InputAndDeleteButton.style.ts
@@ -1,0 +1,9 @@
+import {css} from '@emotion/react';
+
+export const InputAndDeleteButtonContainer = () =>
+  css({
+    display: 'flex',
+    alignItems: 'center',
+    width: '100%',
+    gap: '1rem',
+  });
